backend/requirements.txt
google-genai==1.70.0
boto3==1.42.84
Pillow==12.2.0
pydantic==2.12.5
pydantic-core==2.41.5
